classified analyze_frames for better or worse
Former-commit-id: 6aee256f20475ec60b01d8b6b9a469a3e04751d7
