+++ GestureSentenceManager.py
@@ -1,0 +1,11 @@
+
+
+
+## manages gesture and sentence stuff.
+
+
+
+
+class GestureSentenceManager():
+    def __init__(self, gesture_data, seeds):
+        self.Clusterer =
